MB-33930 [SR]: Enable PassiveDM at vbstate-pending
We need a working instance of PassiveDM for VBuckets in vbstate-pending.

Two reasons for that:
1) ns_server can potentially switch to/from vbstate replica/pending at
  any time (e.g., when rebalance fails)
2) at active takeover, the new active is backfilled while in
  vbstate-pending

In both cases we need to queue incoming Prepares into the PassiveDM for
tracking. In case (2), the PassiveDM is temporary and it will be flipped
to ActiveDM as soon as ns_server issues set-vbucket-state(active) at
takeover completed.

Change-Id: Ie273a0a9f434a2f74052cde1dac6714ae95b3aa3
Reviewed-on: http://review.couchbase.org/108458
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>