MB-33959: Refactor SyncWrite for a second replicationChain
a) Replace SyncWrite members with ReplicationChain*.

Many members can be replaced with a pointer (ideally a reference
but Passive DM does not have ReplicationChains) to a
ReplicationChain. This makes it easier to add the second chain
as we do not need to duplicate majority and active in each SyncWrite
for the second ReplicationChain.

b) Rename variables/output strings to indicate they belong to the
first chain.

Change-Id: If4564131b7ade7dcdfd5217df20ebf3ca6787867
Reviewed-on: http://review.couchbase.org/108460
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>