MB-33959: Add replica tracking for ADM::State::secondChain
Add the tracking of replica acks for the secondChain in the ADM.
Modify SyncWrite::isSatisfied() to also require that the secondChain
meets the durability requirements.

Also, only commit SyncWrite if new active acks

If we are rebalancing an active vBucket to a new node, then we need
to ensure that SyncWrites are acked by the new active. This should be
done by checking that the secondChain active acks if we have a
secondChain (rebalancing) and the secondChain active differs to the
firstChain.

Currently a SyncWrite may time out if a replica from the second chain
acks before the change to a dual topology is propagated to us.

Change-Id: I72b8a0901c75a5a4c2932d3110ddc6903f119ff4
Reviewed-on: http://review.couchbase.org/108507
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>