MB-33971: System connections is bound to system port
There was a misunderstanding when the task to account for system
connections. I thought we wanted to account for all "couchbase
internal components", but ns_server have no control over the
connection usage for components like eventing, indexing, projector
query, fts etc, and we don't want to end up in a situation where
ns_server can't connect to the cluster because the pool of system
connections is consumed.

To work around this ns_server will mark an interface as "system"
and all connections towards that port is system connections.

Change-Id: I65384db1411475c7c76f7a7182931dbe9e0e680f
Reviewed-on: http://review.couchbase.org/108527
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>