MB-32591: Swap instead of moving the dummy item
When creating the list of items to expel (expelItems), we were moving
the item pointed to by the cursor (i.e. the last processed item) into
expelItems and then setting the entry in the list to point to the dummy.

This is shown in the following example:

checkpoint list
===============

[0] -> dummy
[1] -> checkpoint start
[2] -> Item0
[3] -> Item1 <<<<<<< cursor to expel from
[4] -> Item2

1. Move Item1 to expelItems list.
2. splice from dummy to Item1, which will transfer dummy, checkpoint
start and Item0 to expelItems.  The important point is that the dummy
item will now have 2 references to it (one from the orignal checkpoint
list and one from the expelItems list.)

So the lists will now look as follows:

checkpoint list                   expel list
===============                   ==========
[0] -> dummy <-----------------   [0]
[1] -> Item2                      [1] -> checkpoint start
                                  [2] -> Item0
                                  [3] -> Item1

This means expelItems list contains 4 entries which makes a proposed
future patch for calculating the memory saved through expelling,
complex to implement.

Therefore this patch uses swap instead of moving the dummy item.
With this change the lists after expelling look as follows:

checkpoint list                   expel list
===============                   ==========
[0] -> dummy                      [0] -> checkpoint start
[1] -> Item2                      [1] -> Item0
                                  [2] -> Item1

Change-Id: Iacb8dafcf0d55e4f93c7828947bbd65d7a06d94d
Reviewed-on: http://review.couchbase.org/108581
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>