MB-33396: Return ENGINE_DURABILITY_IMPOSSIBLE
For supported operations, return ENGINE_DURABILITY_IMPOSSIBLE
before performing any action (e.g putting something in the hash table)
if the operation is a SyncWrite and durability is not possible with
the current replication topology.

Missing get_and_touch as currently ep_engine support for durability does
not extend to get_and_touch.

Change-Id: Ib17cceb310c2cc846ebfe1504095ec768dd66b10
Reviewed-on: http://review.couchbase.org/108771
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>