MB-34027 [SR]: Implement default timeout
For SyncWrite requests which don't specify a timeout, a
server-provided timeout should be applied to the request.

Add support for this - when a SyncWrite is added to the
ActiveDurabiltyMonitor, if the requests' durability requirements
specify a default timeout then apply a value - currently 30s.

Add a new cb::durability::Timeout class to implement this, which
enforces the two special values the timeout can have:

- BucketDefault: Use the default timeout value the bucket is
  configured with.
- Infinite: Don't use any timeout value (wait forever).

Note that Infinite is used internally (for example after Warmup /
replica promotion were we _must_ wait for prepared SyncWrites to
complete), but is not valid for clients to specify.

Similary BucketDefault should only be specified by a client via
omitting the timeout field - an explcit timeout value of '0' is
invalid.

Change-Id: Ic4850f35b44307ccadba89145c8a4a589b089754
Reviewed-on: http://review.couchbase.org/108656
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>