Reduce #includes of durability_spec.h
durability_spec.h is included by protocol_binary.h, which results in
444 targets (!) needing to be recompiled every time durability_spec.h
is changed.

Instead, forward-declare cb::durability::Requirements in
protocol_binary.h, and then add appropriate #includes of
durability_spec.h to just those targets which need it.

This reduces the number of targets which must be rebuilt after a
change in durability_spec.h to 285. This is still a large number but a
significant decrease.

Change-Id: I8164d74f8954c16ef96ae1794f69fdcee567c135
Reviewed-on: http://review.couchbase.org/108655
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>