MB-32591: Update getActiveVBucketsSortedByChkMgrMem to not filter
Remove "Active" from the name of getActiveVBucketsSortedByChkMgrMem. so
it becomes getVBucketsSortedByChkMgrMem.

Modify the function that it no longers filters on the state of vbuckets
but instead considers all vbuckets.  This enables the function to not
only return active vbuckets, but also replica vbuckets.

This functionality will be needed for when we trigger checkpoint
expelling using the same trigger that it currently used for cursor
dropping.

Change-Id: Ic4936934223f7dc0104f46338538d5ba7bb478a1
Reviewed-on: http://review.couchbase.org/108672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>