MB-32591: Update getVBucketsTotalCheckpointMemoryUsage to not filter
Remove "Active" from the name of getActiveVBucketsTotalCheckpointMemoryUsage
so it becomes getVBucketsTotalCheckpointMemoryUsage.

Modify the function that it no longers filters on the state of vbuckets
but instead considers all vbuckets.  This enables the function to
return the total checkpoint memory usage of all vbuckets, as opposed to
only active vbuckets.

This functionality will be needed for when we trigger checkpoint
expelling using the same trigger that it currently used for cursor
dropping.

Change-Id: I8cec5dc3d90f86cdb89d5a96b66919cf3ca31cda
Reviewed-on: http://review.couchbase.org/108760
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>