MB-34092: Modify doc counts for full eviction
Fix an issue in doc counting for full eviction where we subtracted
the number of pending SyncWrites from the number of items on disk.
This was incorrect as the number of items on disk does not include the
number of pending SyncWrites in the first place so can lead to incorrect
doc counts (e.g. 1 on disk committed item + 1 prepare on that item =
0 items returned from EPVBucket::getNumItems()).

This change to doc counting causes the issue described in MB-34092
where full eviction doc counts are now incorrect post warmup.

This change was tested by enabling full eviction variants of durability
tests. These tests will be enabled in master in the following commit to
highlight this change/issue.

Change-Id: I7bed9b7cfa5adc3e71aaea481da4cd7bd184887b
Reviewed-on: http://review.couchbase.org/108842
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>