MB-33984 [SR]: Maintain ActiveDM state at active -> active transition
When changing vbucket topology but _not_ changing state (i.e. state
stays as active), do not re-create the ActiveDM, keep the existing one
and just apply the new topology.

This also removes the prevVBState parameter from
setupSyncReplication() - we don't require it, we can simply check the
current DurabilityMonitor object type (which is less error-prone than
trying to keep a state and object type in sync).

Change-Id: I4705eda06736adc6bd0bebc99775812e06497852
Reviewed-on: http://review.couchbase.org/108894
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>