MB-29117: Consumer::step should return EWOULDBLOCK until add_stream
ns_server is not prepared to receive DCP control messages before
it sent the ADD STREAM to the DCP consumer. The core did not treat
the connection as a DCP connection until it either received ADD STREAM
or STREAM REQ, so we would not end up trying to send the messages.

We do however want to simplify the logic in the core as a preparation
for moving to bufferevents and out of order execution. DCP connections
should be treated separately in the core, so we should mark them as
DCP immediately when we receive DCP OPEN (separate patch)

Change-Id: Iaedf8c2f305a3d239ee7b9131291410b903c0287
Reviewed-on: http://review.couchbase.org/108935
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>