MB-34027 [SR]: Hardcode infinite timeout for DCP_PREPARE
DCP_PREPAREs currently encode the durability timeout as part of the
durability requirements. This is sent by the producer as an Infinite
value (0xffff), which is logically correct as once a Prepare has been
sent to a replica, _if_ the replica was to be promoted we'd have to
wait forever (it cannot abort).

However, since commit 7240e771e it is invalid to use either of the
reserved values 0x0 or 0xffff for a cb::durability::Level object -
doing so results in DCP_PREPAREs being rejected by the replica:

    2019-05-09T15:59:18.724045+01:00 WARNING 47: exception occurred in runloop during packet execution.
    ... "opcode":"DCP_PREPARE" ...
    Timeout(): Cannot specify an infinite timeout

Given this is breaking _all_ SyncWrites, make a minimal change to
ignore what timeout has been encoded on the wire for DCP_PREPARE and
just hardcode Infinity().

A more complete / cleaner solution (to follow) would be remove the
timeout field from DCP_PREPARE given it's unnecessary to encode it on
the wire.

Change-Id: Iad09077425b86bcbcdf226fca7c10e152471386b
Reviewed-on: http://review.couchbase.org/108902
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>