MB-34121 [SR]: Remove unused DCP_COMMIT.prepare_seqno
Changes in the DCP protocol encoding mean that we are now using the
key of a prepared SyncWrite to identify it when performing a
DCP_COMMIT.

As such, the prepared_seqno field in DCP_COMMIT is redundant and not
required.

Note this field was already ignored the VBucket::commit() - note
pendingCommit argument is unused.

Change-Id: I34468dfb44ce40335c01bb7d0518e779152cd3c4
Reviewed-on: http://review.couchbase.org/108945
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>