MB-34129: CouchRocks: Disable per-CF memory quotas
When running with RocksDB's debug assertions enabled
(i.e. Debug/RelWithDebInfo build), the following assertion fires when
running ep-engine_ep_unit_tests:

    [----------] 8 tests from AllBucketTypes/DurabilityActiveStreamTest
    ...
    [ RUN      ] AllBucketTypes/DurabilityActiveStreamTest.SendDcpPrepare/persistentRocksdb_value_only
    Assertion failed: (!ShouldScheduleFlush()), function MemTable, file ../db/memtable.cc, line 110.

Investigation of that code in RocksDB suggests that the MemTable
buffer sizes / used memory break internal expectations. I don't
understand the exact problem, but it seems to be related to us trying
to set the MemTable write_buffer_size too small; possibily also
related to doing this after the ColumnFamily has been created.

Note that RocksDBKVStore does adjust the MemTable quotas for the
default and seqno column families after a vBucket is created, in
applyMemtablesQuota() - see http://review.couchbase.org/88360.

Disabling this per-VB/CF adjustment, and just setting an overall limit
for the shard solves the problem. This is not necessarily ideal, as
previous analysis found it useful to give the default CF (which stores
key+values) a greater percentage than the seqno tree.

Change-Id: I8d83d71ddd6857ac2d0676f5880d6fd56e29fcda
Reviewed-on: http://review.couchbase.org/108963
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>