Add cluster test framework
The cluster test framework allows you to create a cluster of 'n' nodes
on the same machine to test cross-node functionality (replication,
durability etc).

    auto cluster = cb::test::createCluster(4);
    auto bucket = cluster->createBucket("test",
                                        {{"replicas", 3},
                                         {"max_size", 67108864}});

At this point you've got a cluster with a bucket named test and all of
the replication streams should be up'n'running. The cluster provides
a way to get a connection to a given node, or to get a connection to
the node where a given vbucket lives (with the specified state).
Note that you would need to authenticate and select bucket (the system
can't select the bucket for you as that would need to happen after
an authentication phase):

    auto connection = bucket->getConnection(Vbid{0});
    connection->authenticate("foo", "bar", "PLAIN");
    connection->selectBucket("test");
    conn->store("foo", Vbid(0), "value");

The test program does not work on windows yet due to:

1. process startup and teardown (I didn't bother implement it as I
   don't have the setup to test it right now. I'll spin up a VM
   later to fix it)
2. Program path's (I've not sanitized them)

Change-Id: I9d626adc2fff065a07992754f2ae0e33818043cf
Reviewed-on: http://review.couchbase.org/109100
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>