Refactor: Add StoredValueProxy to HashTable
Commit and abort for SyncWrites should not be functions on
HashTable as the implementation differs greatly between EP and
Ephemeral buckets. We could implement each function by manually
setting StoredValue&'s, but this leaves the HashTable stats in a
bad state. It's too error prone to rely on the callers to set the
stats correctly, so add the StoredValueProxy (FindResult
alternative) to give HashTable users finer control of the
StoredValues without breaking stats.

Change-Id: I1f98c681ddfada1be51363909126bf8dff7bc3db
Reviewed-on: http://review.couchbase.org/109040
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>