Add HashTable delete by StoredValue ptr
Add a function to the HashTable to perform a delete of the given
StoredValue in the given HashBucket. This is useful when we know
which StoredValue we wish to delete but there could be multiple
per key i.e. a committed and pending SyncWrite. It's more flexible to
simply compare a pointer than pass through some given state.

Change-Id: I8400d6cfaf70a18392d380997a74031f4060e608
Reviewed-on: http://review.couchbase.org/109041
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>