MB-34017 [SR]: Implement the High Completed Seqno (HCS) in ActiveDM
The High Completed Seqno is the sequence number of the highest prepared
SyncWrite that has been completed, i.e.:

- whether the Prepare has globally met its durability requirements and
  has been committed by the active node
- or, timeout has triggered on Active for the Prepare and it has been
  aborted
- and (in either cases) all earlier SyncWrites have been completed.

The HCS at Active is going to be used as an optimization at Warmup.

Currently, at Warmup we need to perform a first phase where we look for
Prepares that have not been completed (Committed or Aborted) and need
to be re-processed. Given that we don't know what was the last completed
Prepare, we need to scan the entire seqno-index for completing this
Prepare-phase. So, we end-up with scanning the entire seqno-index twice:
once for the Prepare-phase and again for the usual LoadKey+LoadValue
phase. That leads to a 100%-increase on the Warmup runtime.

By introducing the HCS, the Prepare-phase at Warmup can just scan the
seqno-index starting from the HCS, thus avoiding a double full scan.

Change-Id: Icb6920df3c2c0dbd6a7c878c15118d263ef3ff0c
Reviewed-on: http://review.couchbase.org/109115
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>