MB-33332: Move HashTable::abort to EPVBucket
The HashTable::abort function is specific to Persistent vBuckets
and cannot be used for Ephemeral vBuckets. This logic is functionally
the same as a delete of the pending for the given key. We can
accomplish this by calling ht.unlocked_del with a pointer to the
StoredValue that we want to delete.

Change-Id: Id12154c131d478eef3dbc67fbafc1a5caea4aa87
Reviewed-on: http://review.couchbase.org/109120
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>