Avoid unlocked_find() after VBucket::addTempStoredValue()
VBucket::addTempStoredValue() is used to create a temporary item in
the HashTable for subsequent BGfetch. In the majority of use-cases
callers of this then need the newly-created SV; to achieve this they
call HashTable::unlocked_find() on the key.

However this call to unlocked_find() is:

a) Inefficient, given that VBucket::addTempStoredValue() just
   performed unlocked_find() and we are repeating work unnecessarily.

b) Adds additional use-cases for unlocked_find() which cannot be
fulfilled by the higher level HashTable::findFor...() methods,  as the
HashBucketLock has already been acquired (and HashTable::findFor...()
acquire the hbl themselves.

For the above reasons (primarily b), modify
VBucket::addTempStoredValue() to return a ptr to the newly-created
StoredValue, avoiding the need for the extra unlocked_find() calls.

Change-Id: I17716e9329425fd305c53f279b4d73aadcaac450
Reviewed-on: http://review.couchbase.org/109316
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>