Use MockEphemeralVBucket in VBucketDurabilityTest
To add some commit and SyncDelete tests in
the future, we need to be able to manually check seqList stats and
correct the collections stats for Ephemeral to prevent counter
underflow assertions. The underflow occurs because we don't hit the
public functions in VBucketDurabilityTest (we would need an engine to
do so) we miss a call to update the collections stats on a set.

Change-Id: If72a0b97c9b56415d68caec1fadb22e143ac63aa
Reviewed-on: http://review.couchbase.org/109199
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>