Simplify DCP handling in the core
This patch marks the connection as a DCP connection as part of
DCP Open instead of delaying that until the first Add Stream
or Stream Request.

Future refactoring of the core needs this in order to correctly
identify DCP connections as part of notify_io_complete (test
that the cookie is in a blocked state when it is notified, and
that's not true for DCP connections).

Change-Id: I67aedeceec1609b6edbc587aaada9b7b0e2576d7
Reviewed-on: http://review.couchbase.org/108786
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>