Avoid unlocked_find() in VBucket::fetchValidValue
Remove the use of HashTable::unlocked_find() from
VBucket::fetchValidValue(), replacing it with HashTable::findForRead().

Note this includes a somewhat undesirable cast away of const-ness on
the findForRead() result. This is needed because unfortunately
VBucket::fetchValidValue does need a non-const SV to be able to
perform expiry, and also because various existing callers of this
method expect a non-const SV.

The least-worst option (IMHO) is to have HashTable::findForRead()
return a const SV so at least other use-cases _can_ be const-correct,
even if we cannot be here.

Ideally this should be removed; but that requires that findForRead()
is refactored to return a proxy object which while logically const,
*does* allow some "physically" const methods (like
ht.unlocked_restore) to be called on it. That's a non-trivial change
to defer for now.

Change-Id: I926188e36084e8c8cbc06e360ac69ee1ad315efd
Reviewed-on: http://review.couchbase.org/109319
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>