Add HashTable::findOnly{Committed,Prepared}
While in most cases HashTable::findFor{Read,Write} should be used to
return the correct "client" view of the HashTable, there are some
instances where the caller *needs* access to a only the Comitted or
Prepared StoredValues.

These cases are typically non-frontend instances - for example
completing BGFetches or Persistence callbacks.

As such, add two methods to HashTable to facilitate this:

- findOnlyCommitted - Looks for a Commited item with the given key,
  returns nullptr if not exists; ignoring any Prepared SVs.

- findOnlyPrepared - Looks for a Prepared item with the given key,
  return nullptr if not exists; ignoring any Committed SVs.

Note these functions should be used with care - if you are considering
using this to implement a front-end operation then consider if
findForRead() / findForWrite() should be used instead.

Change-Id: If52080d283fa24cf185a36ffdd466526daac8fc5
Reviewed-on: http://review.couchbase.org/109334
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>