MB-34228: Fix deadlock for SyncWrite with Replicas=0
There exists a deadlock issue for SyncWrites when Replicas=0 where
we attempt to acquire the HashBucketLock again whilst we already have
it acquired. This is because we attempt to immediately commit a
SyncWrite if it is satisfied.

Update the ADM::addSyncWrite function to not attempt to commit the
SyncWrite immediately. Instead, we will commit the SyncWrite at the
end of the VBucket::set et al functions after we no longer need the
HashBucketLock (although we will re-acquire the HashBucketLock if
we commit).

Add a test for SyncWrite with Replicas=0.

Change-Id: Ied14a712bf735ad9b0f93955e0e336e6dd50080f
Reviewed-on: http://review.couchbase.org/109360
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>