Add support for auto retry for etmpfail
The current code use a hardcoded backoff with 10ms and 1 minute
timeout.

Note that the primary intention with this client is to be able
to test the server so we want to be able to read out all error
codes returned by the server (and not be a full fledged xerror
capable client).

Change-Id: I4d2aa5094004aaf8499f1f995a22703f336cd1a3
Reviewed-on: http://review.couchbase.org/109421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>