MB-33958 [SR]: 1/3 Add CommittedState::PreparedMaybeVisible
Add a second variant of CommittedState::Pending -
CommittedState::PreparedMaybeVisible:

Item is prepared, but *may* have already been committed by another
node or before a restart, and as such cannot allow access to *any*
value for this key until the SyncWrite is committed.  Same semantics
as 'Pending, with the addition of blocking reads to any existing
value.

This patch adds the enumeration value to StoredValue and
Item. Subsequent patches will make use of it, setting warmed up
Prepares to the new state and blocking reads if in this state.

Change-Id: I9bad8073ca21e861002c2813acd9b91ab75807c0
Reviewed-on: http://review.couchbase.org/109444
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>