MB-33958 [SR]: 3/3 Flag warmed-up prepared SyncWrites as MaybeVisible
For prepared SyncWrites which are loaded during Warmup, we cannot
allow any client access to that document:

- If we returned the old value (prior to the prepare), then we break
  serializability, as the client (pre-warmup) could have already
  observed a success message for the "committed" prepare (but the
  commit wasn't persisted to disk before restart).

- If we return the new (not yet committed value) we have the opposite
  problem, as that mutation hasn’t yet been made sufficiently durable
  to be able to commit.

To correctly handle this situation, mark all Prepared SyncWrites
loaded during warmup as "PreparedMaybeVisible" - they *may* have
already been made visible to clients, and as such until they are
re-committed we must not return a value for them.

Update HashTable::findForRead() to return the prepared SV for a key if
that Prepare is MaybeVisible; callers of this are updated to return
the correct status code to the clients.

This also requires that the subsequent warmup of Committed items uses
the HashTable::findOnlyCommitted() method so it doesn't get redirected
to the (already warmed-up) prepared item.

Change-Id: I949c18f9639f75a09c3288d29b8478621c5a4bd5
Reviewed-on: http://review.couchbase.org/109446
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>