MB-33951 [SR]: Flag PDM->ADM Prepares as MaybeVisible (re-apply)
(re-apply after revert).

For prepared SyncWrites existed in the Passive DM when it was
converted to Active DM, we cannot allow any client access to that
document:

- If we returned the old value (prior to the prepare), then we break
  seriailzability, as the client (pre-warmup) could have already
  observed a success message for the "committed" prepare (but the
  commit wasn't persisted to disk before restart).

- If we return the new (not yet committed value) we have the opposite
  problem, as that mutation hasn’t yet been made sufficiently durable
  to be able to commit.

Note that given that in a PassiveDM all Prepares will either:

a) Complete (be Committed/Aborted) via DCP in which case they will be
   removed from the DM / HashTable, or:

b) Be moved into an Active Durabilty Monitor if the vBucket is
   promoted to active - in which case we want all such prepares to be
   marked as MaybeVisible.

Therefore we can simply _always_ create Prepared items on a non-active
node (DCP Consumers) as PreparedMaybeVisible, and don't have to
convert any incomplete Prepares if the vBucket state changes to
active.

Change-Id: I773cc4e404e75da365d204cfddbe73d64c068836
Reviewed-on: http://review.couchbase.org/109561
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>