MB-34151 [SR]: Remove Prepares from PassiveDM only if locally-satisfied
Differently from the ActiveDM, Prepares in the PassiveDM can be
completed (Committed/Aborted) *before* they are locally-satisfied on the
Replica node. That may be the case for Level PersistToMajority, as the
PassiveDM may receive a Commit/Abort before the Prepare is persisted
locally.

Given that the move-logic of HPS relies on the content of the tracked
Container, the above leads to wrong (or missing) HPS updates in the case
where the new position for HPS (logically) points to a Prepare that has
been removed from tracking.

To fix, we don't remove completed Prepares if they are not
locally-satisfied. I.e., a Prepare is eligible for removal:
1) if it is completed
2) and, if it is locally-satisfied

For achieving (1), this change introduces a new quantity in PassiveDM:
the High Completed Seqno. The HCS points to the last completed Prepare
(ie, Committed or Aborted).

Now a Prepare is *not* removed from tracking unconditionally as soon as
it is notified for completion. At completion, a Prepare is just marked
completed by moving the HCP on to it.

Then, every time the HCP or the HPS move we check for new Prepares
eligible for removal, and we remove them if any.

Change-Id: Ie10eaf35e37859422b42fe26df12b2eee476ea98
Reviewed-on: http://review.couchbase.org/109452
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>