MB-34094,MB-34120: Fix on disk stats underflow and add unit tests
Fix underflow of onDiskTotalItems in EPVBucket seen in MB-34120 and an
underflow of diskCount in ManifestEntry which stores the number of items
on disk for a given collection, which was observed in MB-34094.

This bug is due to us not correctly incrementing
EPVBucket::onDiskTotalItems every time insert a new mutation on disk,
which should include prepares. As currently we do not treat sets to
prepared and committed documents as being  separate documents. We need
Do do this as couchstore and our other storage engines have no concept
of the relationship between a prepared and a committed document.

To fix this, we've modified the set PersistenceCallback to increment
onDiskTotalItems for new prepare sets as well as a new key commits.
Also, when looking up the persisted item from the HashTable
during the set or delete (see VBucket::deletedOnDiskCbk())
PersistenceCallback's, we search for the correct pending / committed
document using HashTable::findOnlyPrepared() /
HashTable::findOnlyCommitted() respectively depending on the queued
items context. Instead of using fetchValidValue which returns
Committed or Pending depending on the state of the HashTable, meaning
we could be in a callback context for a prepare but end up treating it
as a commit.
In some contexts findOnlyPrepared() dose not return a StoredValue.
However, if the queued time is pending and the mutation status passed
to the set persistence callback is
KVStore::MutationSetResultState::Insert then we increment the on item
count. As this implies the item isn't available in the HashTable right
now but we have just written a new prepare to disk.

This patch also adds a set of unit tests to verify that our accounting
of the number of documents on disk are correct.

Change-Id: I4e6933aabf7b8424f08a6d1126a16683cc7849c9
Reviewed-on: http://review.couchbase.org/109642
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>