MB-34091: Manually ack seqnos for new secondChain nodes
It is possible for a new node (will exist in the new topology)
to ack before ns_server gives us a new replication topology.
ns_server does this so that we do not block SyncWrites on vBucket
streaming on the new node as this could take a long time and cause
the SyncWrites to time out. However, this means that a new node can
ack before it exists in our replication topology.

Store a map of acks for nodes that we do not know about, and use this
map on setting of the replication topology to manually ack each new
node at their highest acked seqno. This will ensure that our SyncWrites
do not timeout waiting for the secondChain to be satisfied.

Change-Id: I480abf22878b30d321b3ffb4419f61975d33c5eb
Reviewed-on: http://review.couchbase.org/109575
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>