MB-34262: Replicate ephemeral tombstone times
1) Ephemeral backfill wasn't copying the tombstone time from
the OSV deleted time field into the outgoing Item, this DCP
delete v2 sent 0 or the expiry time to the client. This is
corrected in the memory backfill so that the time is now
passed to the Item that DCP will use to build the outbound
message.

2) The ephemeral DCP consumer path wasn't doing the reverse of 1 (same for
DelWithMeta). That is when the replicated delete is pushed into the HT/seqno
linked list, we didn't copy through any delete time, a new delete time was
always being generated.

Change-Id: I7457acd699766a1a029d636663f50aebb479d934
Reviewed-on: http://review.couchbase.org/109560
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>