MB-34316 [SR]: Use VBucket::fetchValueForWrite for lock/unlock/GAT
Add a new method to VBucket::fetchValueForWrite() which is a sibling
to the existing fetchValidValue().

fetchValueForWrite() is intended when the caller wants to acquire an
item to perform a modification - with the advent of prepared items for
SyncWrites, the caller needs to speficy their intent to be directed to
the correct StoredValue.

Update existing callers of fetchValidValue() which were incorrect -
they were actually fetching an item with write intent, which is
incorrect to use fetchValidValue for.

Change-Id: Id543121ce2908914ffe733c79e348e1fd6d4a1bb
Reviewed-on: http://review.couchbase.org/109567
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>